+++ Gruntfile.js
@@ -1,5 +1,21 @@
 'use strict';
 module.exports = function(grunt) {
+  var jsFileList = [
+    'assets/vendor/bootstrap/js/transition.js',
+    'assets/vendor/bootstrap/js/alert.js',
+    'assets/vendor/bootstrap/js/button.js',
+    'assets/vendor/bootstrap/js/carousel.js',
+    'assets/vendor/bootstrap/js/collapse.js',
+    'assets/vendor/bootstrap/js/dropdown.js',
+    'assets/vendor/bootstrap/js/modal.js',
+    'assets/vendor/bootstrap/js/tooltip.js',
+    'assets/vendor/bootstrap/js/popover.js',
+    'assets/vendor/bootstrap/js/scrollspy.js',
+    'assets/vendor/bootstrap/js/tab.js',
+    'assets/vendor/bootstrap/js/affix.js',
+    'assets/js/plugins/*.js',
+    'assets/js/_*.js'
+  ];
 
   grunt.initConfig({
     jshint: {
@@ -9,75 +25,92 @@ module.exports = function(grunt) {
       all: [
         'Gruntfile.js',
         'assets/js/*.js',
-        '!assets/js/scripts.min.js'
+        '!assets/js/scripts.js',
+        '!assets/**/*.min.*'
       ]
     },
     less: {
-      dist: {
+      dev: {
         files: {
-          'assets/css/main.min.css': [
-            'assets/less/app.less'
+          'assets/css/main.css': [
+            'assets/less/main.less'
           ]
         },
         options: {
-          compress: true,
+          compress: false,
           // LESS source map
           // To enable, set sourceMap to true and update sourceMapRootpath based on your install
           sourceMap: false,
-          sourceMapFilename: 'assets/css/main.min.css.map',
+          sourceMapFilename: 'assets/css/main.css.map',
           sourceMapRootpath: '/app/themes/roots/'
         }
+      },
+      build: {
+        files: {
+          'assets/css/main.min.css': [
+            'assets/less/main.less'
+          ]
+        },
+        options: {
+          compress: true
+        }
       }
+    },
+    concat: {
+      options: {
+        separator: ';',
+      },
+      dist: {
+        src: [jsFileList],
+        dest: 'assets/js/scripts.js',
+      },
     },
     uglify: {
       dist: {
         files: {
-          'assets/js/scripts.min.js': [
-            'assets/js/plugins/bootstrap/transition.js',
-            'assets/js/plugins/bootstrap/alert.js',
-            'assets/js/plugins/bootstrap/button.js',
-            'assets/js/plugins/bootstrap/carousel.js',
-            'assets/js/plugins/bootstrap/collapse.js',
-            'assets/js/plugins/bootstrap/dropdown.js',
-            'assets/js/plugins/bootstrap/modal.js',
-            'assets/js/plugins/bootstrap/tooltip.js',
-            'assets/js/plugins/bootstrap/popover.js',
-            'assets/js/plugins/bootstrap/scrollspy.js',
-            'assets/js/plugins/bootstrap/tab.js',
-            'assets/js/plugins/bootstrap/affix.js',
-            'assets/js/plugins/*.js',
-            'assets/js/_*.js'
-          ]
-        },
-        options: {
-          // JS source map: to enable, uncomment the lines below and update sourceMappingURL based on your install
-          // sourceMap: 'assets/js/scripts.min.js.map',
-          // sourceMappingURL: '/app/themes/roots/assets/js/scripts.min.js.map'
+          'assets/js/scripts.min.js': [jsFileList]
         }
       }
     },
-    version: {
-      options: {
-        file: 'lib/scripts.php',
-        css: 'assets/css/main.min.css',
-        cssHandle: 'roots_main',
-        js: 'assets/js/scripts.min.js',
-        jsHandle: 'roots_scripts'
+    modernizr: {
+      build: {
+        devFile: 'assets/vendor/modernizr/modernizr.js',
+        outputFile: 'assets/js/vendor/modernizr.min.js',
+        files: {
+          'src': [
+            ['assets/js/scripts.min.js'],
+            ['assets/css/main.min.css']
+          ]
+        },
+        uglify: true,
+        parseFiles: true
+      }
+    },
+    filerev: {
+      assets: {
+        src: ['assets/css/main.min.css', 'assets/js/scripts.min.js']
+      }
+    },
+    filerev_assets: {
+      assets: {
+        options: {
+          dest: 'assets-manifest.json'
+        }
       }
     },
     watch: {
       less: {
         files: [
           'assets/less/*.less',
-          'assets/less/bootstrap/*.less'
+          'assets/less/**/*.less'
         ],
-        tasks: ['less', 'version']
+        tasks: ['less']
       },
       js: {
         files: [
           '<%= jshint.all %>'
         ],
-        tasks: ['jshint', 'uglify', 'version']
+        tasks: ['jshint', 'concat']
       },
       livereload: {
         // Browser live reloading
@@ -95,29 +128,32 @@ module.exports = function(grunt) {
     },
     clean: {
       dist: [
-        'assets/css/main.min.css',
-        'assets/js/scripts.min.js'
+        'assets/css/*main*.min.css',
+        'assets/js/*scripts*.min.js'
       ]
     }
   });
 
   // Load tasks
-  grunt.loadNpmTasks('grunt-contrib-clean');
-  grunt.loadNpmTasks('grunt-contrib-jshint');
-  grunt.loadNpmTasks('grunt-contrib-uglify');
-  grunt.loadNpmTasks('grunt-contrib-watch');
-  grunt.loadNpmTasks('grunt-contrib-less');
-  grunt.loadNpmTasks('grunt-wp-version');
+  require('load-grunt-tasks')(grunt);
 
   // Register tasks
   grunt.registerTask('default', [
-    'clean',
-    'less',
-    'uglify',
-    'version'
+    'dev'
   ]);
   grunt.registerTask('dev', [
-    'watch'
+    'jshint',
+    'less:dev',
+    'concat'
+  ]);
+  grunt.registerTask('build', [
+    'jshint',
+    'clean',
+    'less:build',
+    'uglify',
+    'modernizr',
+    'filerev',
+    'filerev_assets'
   ]);
 
 };