+++ lib/scripts.php
@@ -3,22 +3,48 @@
  * Enqueue scripts and stylesheets
  *
  * Enqueue stylesheets in the following order:
- * 1. /theme/assets/css/main.min.css
+ * 1. /theme/assets/css/main.css
  *
  * Enqueue scripts in the following order:
  * 1. jquery-1.11.0.min.js via Google CDN
- * 2. /theme/assets/js/vendor/modernizr-2.7.0.min.js
- * 3. /theme/assets/js/main.min.js (in footer)
+ * 2. /theme/assets/js/vendor/modernizr.min.js
+ * 3. /theme/assets/js/scripts.js (in footer)
  */
 function roots_scripts() {
+
+  /**
+   * The build task in Grunt renames production assets with a hash
+   * Read the asset names from assets-manifest.json
+   */
+  if (!(WP_ENV === 'development')) {
+    $get_assets = file_get_contents(get_template_directory_uri() . '/assets-manifest.json');
+    $assets     = json_decode($get_assets, true);
+    $assets     = array(
+      'css'       => '/' . $assets['assets/css/main.min.css'],
+      'js'        => '/' . $assets['assets/js/scripts.min.js'],
+      'modernizr' => '/assets/js/vendor/modernizr.min.js',
+      'jquery'    => '//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js'
+    );
+  } else {
+    $assets = array(
+      'css'       => '/assets/css/main.css',
+      'js'        => '/assets/js/scripts.js',
+      'modernizr' => '/assets/js/vendor/modernizr.min.js',
+      'jquery'    => '//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.js'
+    );
+  }
+
+  wp_enqueue_style('roots_css', get_template_directory_uri() . $assets['css'], false, null);
   wp_enqueue_style('roots_main', get_template_directory_uri() . '/assets/css/main.min.css', false, '9880649384aea9f1ee166331c0a30daa');
 
-  // jQuery is loaded using the same method from HTML5 Boilerplate:
-  // Grab Google CDN's latest jQuery with a protocol relative URL; fallback to local if offline
-  // It's kept in the header instead of footer to avoid conflicts with plugins.
+  /**
+   * jQuery is loaded using the same method from HTML5 Boilerplate:
+   * Grab Google CDN's latest jQuery with a protocol relative URL; fallback to local if offline
+   * It's kept in the header instead of footer to avoid conflicts with plugins.
+   */
   if (!is_admin() && current_theme_supports('jquery-cdn')) {
     wp_deregister_script('jquery');
-    wp_register_script('jquery', '//ajax.googleapis.com/ajax/libs/jquery/1.11.0/jquery.min.js', array(), null, false);
+    wp_register_script('jquery', $assets['jquery'], array(), null, false);
     add_filter('script_loader_src', 'roots_jquery_local_fallback', 10, 2);
   }
 
@@ -26,11 +52,9 @@ function roots_scripts() {
     wp_enqueue_script('comment-reply');
   }
 
-  wp_register_script('modernizr', get_template_directory_uri() . '/assets/js/vendor/modernizr-2.7.0.min.js', array(), null, false);
-  wp_register_script('roots_scripts', get_template_directory_uri() . '/assets/js/scripts.min.js', array(), '0fc6af96786d8f267c8686338a34cd38', true);
-  wp_enqueue_script('modernizr');
+  wp_enqueue_script('modernizr', get_template_directory_uri() . $assets['modernizr'], array(), null, false);
   wp_enqueue_script('jquery');
-  wp_enqueue_script('roots_scripts');
+  wp_enqueue_script('roots_js', get_template_directory_uri() . $assets['js'], array(), null, true);
 }
 add_action('wp_enqueue_scripts', 'roots_scripts', 100);
 
@@ -39,7 +63,7 @@ function roots_jquery_local_fallback($src, $handle = null) {
   static $add_jquery_fallback = false;
 
   if ($add_jquery_fallback) {
-    echo '<script>window.jQuery || document.write(\'<script src="' . get_template_directory_uri() . '/assets/js/vendor/jquery-1.11.0.min.js"><\/script>\')</script>' . "\n";
+    echo '<script>window.jQuery || document.write(\'<script src="' . get_template_directory_uri() . '/assets/vendor/jquery/jquery.min.js"><\/script>\')</script>' . "\n";
     $add_jquery_fallback = false;
   }