+++ templates/sidebar.php
@@ -1,1 +1,3 @@
+<?php roots_sidebar_top(); ?>
 <?php dynamic_sidebar('sidebar-primary'); ?>
+<?php roots_sidebar_bottom(); ?>