+++ templates/widgetized-home.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * Template Name: Widgtized Homepage
+ */
+?>
+<?php roots_sidebar_top(); ?>
+<?php dynamic_sidebar('sidebar-home'); ?>
+<?php roots_sidebar_bottom(); ?>